Export and submodules refactored
Refs: https://github.com/metarhia/metasync/issues/176
