+++ lib/array.js
@@ -1,258 +1,267 @@
 'use strict';
 
-module.exports = (api) => {
+const common = require('metarhia-common');
 
-  api.metasync.map = (
-    // Asynchronous map (iterate parallel)
-    items, // incoming array
-    fn, // (current, callback) => callback(err, value)
-    // to be executed for each value in the array
-    //   current - current element being processed in the array
-    //   callback - function(err, value)
-    done // optional on done callback function(err, result)
-  ) => {
-    const len = items.length;
-    const result = new Array(len);
-    let errored = false;
-    let count = 0;
-    done = api.common.cb(done);
-    if (!len) done(null, result);
+const map = (
+  // Asynchronous map (iterate parallel)
+  items, // incoming array
+  fn, // (current, callback) => callback(err, value)
+  // to be executed for each value in the array
+  //   current - current element being processed in the array
+  //   callback - function(err, value)
+  done // optional on done callback function(err, result)
+) => {
+  const len = items.length;
+  const result = new Array(len);
+  let errored = false;
+  let count = 0;
+  done = common.cb(done);
+  if (!len) done(null, result);
 
-    const next = (index, err, value) => {
-      if (errored) return;
-      if (err) {
-        errored = true;
-        return done(err);
-      }
-      result[index] = value;
-      count++;
-      if (count === len) done(null, result);
-    };
-
-    let i;
-    for (i = 0; i < len; i++) {
-      fn(items[i], next.bind(null, i));
+  const next = (index, err, value) => {
+    if (errored) return;
+    if (err) {
+      errored = true;
+      return done(err);
     }
+    result[index] = value;
+    count++;
+    if (count === len) done(null, result);
   };
 
-  api.metasync.filter = (
-    // Asynchrous filter (iterate parallel)
-    items, // incoming array
-    fn, // (value, callback) => (err, accepted)
-    // to be executed for each value in the array
-    //   value - item from items array
-    //   callback - function(err, accepted)
-    done // optional on done callback function(err, result)
-  ) => {
-    const len = items.length;
-    let result = [];
-    let count = 0;
-    let errored = false;
-    done = api.common.cb(done);
+  let i;
+  for (i = 0; i < len; i++) {
+    fn(items[i], next.bind(null, i));
+  }
+};
 
-    if (!len) done(null, result);
+const filter = (
+  // Asynchrous filter (iterate parallel)
+  items, // incoming array
+  fn, // (value, callback) => (err, accepted)
+  // to be executed for each value in the array
+  //   value - item from items array
+  //   callback - function(err, accepted)
+  done // optional on done callback function(err, result)
+) => {
+  const len = items.length;
+  let result = [];
+  let count = 0;
+  let errored = false;
+  done = common.cb(done);
 
-    function finish() {
-      // Callbacks might be called in any possible order,
-      // hence sort the filtered array
-      // by element's index in the original itemsection
-      result.sort((x, y) => (x.index - y.index));
+  if (!len) done(null, result);
 
-      // Only value is needed in resulting array
-      result = result.map(x => x.value);
+  function finish() {
+    // Callbacks might be called in any possible order,
+    // hence sort the filtered array
+    // by element's index in the original itemsection
+    result.sort((x, y) => (x.index - y.index));
 
-      // Return a result using callback;
-      done(null, result);
+    // Only value is needed in resulting array
+    result = result.map(x => x.value);
+
+    // Return a result using callback;
+    done(null, result);
+  }
+
+  const next = (index, err, accepted) => {
+    if (errored) return;
+    if (err) {
+      errored = true;
+      return done(err);
     }
-
-    const next = (index, err, accepted) => {
-      if (errored) return;
-      if (err) {
-        errored = true;
-        return done(err);
-      }
-      if (accepted) result.push({ index, value: items[index] });
-      count++;
-      if (count === len) finish();
-    };
-
-    let i;
-    for (i = 0; i < len; i++) {
-      fn(items[i], next.bind(null, i));
-    }
+    if (accepted) result.push({ index, value: items[index] });
+    count++;
+    if (count === len) finish();
   };
 
-  api.metasync.reduce = (
-    // Asynchronous reduce
-    items, // incoming array
-    fn, // function to be executed for each value in array
-    //   previous - value previously returned in the last iteration
-    //   current - current element being processed in the array
-    //   callback - callback for returning value back to function reduce
-    //   counter - index of the current element being processed in array
-    //   items - the array reduce was called upon
-    done, // optional on done callback function(err, result)
-    initial // optional value to be used as first arpument in first iteration
-  ) => {
-    const len = items.length;
-    done = api.common.cb(done);
-    let count = typeof(initial) === 'undefined' ? 1 : 0;
+  let i;
+  for (i = 0; i < len; i++) {
+    fn(items[i], next.bind(null, i));
+  }
+};
 
-    if (!len) {
-      if (count) {
-        done(new TypeError('Reduce of empty array with no initial value'));
-      } else {
-        done(null, initial);
-      }
-      return;
+const reduce = (
+  // Asynchronous reduce
+  items, // incoming array
+  fn, // function to be executed for each value in array
+  //   previous - value previously returned in the last iteration
+  //   current - current element being processed in the array
+  //   callback - callback for returning value back to function reduce
+  //   counter - index of the current element being processed in array
+  //   items - the array reduce was called upon
+  done, // optional on done callback function(err, result)
+  initial // optional value to be used as first arpument in first iteration
+) => {
+  const len = items.length;
+  done = common.cb(done);
+  let count = typeof(initial) === 'undefined' ? 1 : 0;
+
+  if (!len) {
+    if (count) {
+      done(new TypeError('Reduce of empty array with no initial value'));
+    } else {
+      done(null, initial);
     }
+    return;
+  }
 
-    let previous = count === 1 ? items[0] : initial;
-    let current = items[count];
+  let previous = count === 1 ? items[0] : initial;
+  let current = items[count];
 
-    function response(err, data) {
-      if (err) return done(err);
-      if (count === len - 1) return done(null, data);
-      count++;
-      previous = data;
-      current = items[count];
-      fn(previous, current, response, count, items);
-    }
-
+  function response(err, data) {
+    if (err) return done(err);
+    if (count === len - 1) return done(null, data);
+    count++;
+    previous = data;
+    current = items[count];
     fn(previous, current, response, count, items);
-  };
+  }
 
-  api.metasync.each = (
-    // Asynchronous each (iterate in parallel)
-    items, // incoming array
-    fn, // (value, callback) => callback(err)
-    //   value - item from items array
-    //   callback - callback function(err)
-    done // optional on done callback function(err)
-  ) => {
-    const len = items.length;
-    let count = 0;
-    let errored = false;
-    done = api.common.cb(done);
-    if (len === 0) return done(null);
+  fn(previous, current, response, count, items);
+};
 
-    const next = (err) => {
-      if (errored) return;
-      if (err) {
-        errored = true;
-        done(err);
-      } else {
-        count++;
-        if (count === len) done(null);
-      }
-    };
+const each = (
+  // Asynchronous each (iterate in parallel)
+  items, // incoming array
+  fn, // (value, callback) => callback(err)
+  //   value - item from items array
+  //   callback - callback function(err)
+  done // optional on done callback function(err)
+) => {
+  const len = items.length;
+  let count = 0;
+  let errored = false;
+  done = common.cb(done);
+  if (len === 0) return done(null);
 
-    let i;
-    for (i = 0; i < len; i++) {
-      fn(items[i], next);
+  const next = (err) => {
+    if (errored) return;
+    if (err) {
+      errored = true;
+      done(err);
+    } else {
+      count++;
+      if (count === len) done(null);
     }
   };
 
-  api.metasync.series = (
-    // Asynchronous series
-    items, // incoming array
-    fn, // (value, callback) => callback(err)
-    //   value - item from items array
-    //   callback - callback (err)
-    done // optional on done callback (err)
-  ) => {
-    const len = items.length;
-    let i = -1;
-    done = api.common.cb(done);
+  let i;
+  for (i = 0; i < len; i++) {
+    fn(items[i], next);
+  }
+};
 
-    function next() {
-      i++;
-      if (i === len) return done(null);
-      fn(items[i], (err) => {
-        if (err) return done(err);
-        setImmediate(next);
-      });
-    }
+const series = (
+  // Asynchronous series
+  items, // incoming array
+  fn, // (value, callback) => callback(err)
+  //   value - item from items array
+  //   callback - callback (err)
+  done // optional on done callback (err)
+) => {
+  const len = items.length;
+  let i = -1;
+  done = common.cb(done);
 
-    next();
-  };
-
-  api.metasync.find = (
-    // Asynchronous find (iterate in series)
-    items, // incoming array
-    fn, // (value, callback) => callback(err, accepted)
-    //   value - item from items array
-    //   callback - callback function(err, accepted)
-    done // optional on done callback function(err, result)
-  ) => {
-    const len = items.length;
-    let i = 0;
-    done = api.common.cb(done);
-
-    function next() {
-      if (i === len) return done(null);
-      fn(items[i], (err, accepted) => {
-        if (err) return done(err);
-        if (accepted) return done(null, items[i]);
-        i++;
-        next();
-      });
-    }
-
-    if (len > 0) next();
-    else done(null);
-  };
-
-  api.metasync.every = (
-    // Asynchronous every
-    items, // incoming array
-    fn, // (value, callback) => callback(err, fits)
-    //   value - item from items array
-    //   callback - callback function(err, fits)
-    done // optional on done callback function(err, result)
-  ) => {
-    done = api.common.cb(done);
-    if (items.length === 0) return done(null, true);
-    let proceedItemsCount = 0;
-
-    function finish(err, accepted) {
+  function next() {
+    i++;
+    if (i === len) return done(null);
+    fn(items[i], (err) => {
       if (err) return done(err);
-      if (!accepted) return done(null, false);
+      setImmediate(next);
+    });
+  }
 
-      proceedItemsCount++;
-      if (proceedItemsCount === items.length) done(null, true);
-    }
+  next();
+};
 
-    let item;
-    for (item of items) {
-      fn(item, finish);
-    }
-  };
+const find = (
+  // Asynchronous find (iterate in series)
+  items, // incoming array
+  fn, // (value, callback) => callback(err, accepted)
+  //   value - item from items array
+  //   callback - callback function(err, accepted)
+  done // optional on done callback function(err, result)
+) => {
+  const len = items.length;
+  let i = 0;
+  done = common.cb(done);
 
-  api.metasync.some = (
-    // Asynchronous some (iterate in series)
-    items, // incoming array
-    fn, // (value, callback) => (err, accepted)
-    //   value - item from items array
-    //   callback - callback function(err, accepted)
-    done // on done callback function(err, result)
-  ) => {
-    const len = items.length;
-    let i = 0;
-    done = api.common.cb(done);
+  function next() {
+    if (i === len) return done(null);
+    fn(items[i], (err, accepted) => {
+      if (err) return done(err);
+      if (accepted) return done(null, items[i]);
+      i++;
+      next();
+    });
+  }
 
-    function next() {
-      if (i === len) return done(null, false);
-      fn(items[i], (err, accepted) => {
-        if (err) return done(err);
-        if (accepted) return done(null, true);
-        i++;
-        next();
-      });
-    }
+  if (len > 0) next();
+  else done(null);
+};
 
-    if (len > 0) next();
-    else done(null, false);
-  };
+const every = (
+  // Asynchronous every
+  items, // incoming array
+  fn, // (value, callback) => callback(err, fits)
+  //   value - item from items array
+  //   callback - callback function(err, fits)
+  done // optional on done callback function(err, result)
+) => {
+  done = common.cb(done);
+  if (items.length === 0) return done(null, true);
+  let proceedItemsCount = 0;
 
+  function finish(err, accepted) {
+    if (err) return done(err);
+    if (!accepted) return done(null, false);
+
+    proceedItemsCount++;
+    if (proceedItemsCount === items.length) done(null, true);
+  }
+
+  let item;
+  for (item of items) {
+    fn(item, finish);
+  }
+};
+
+const some = (
+  // Asynchronous some (iterate in series)
+  items, // incoming array
+  fn, // (value, callback) => (err, accepted)
+  //   value - item from items array
+  //   callback - callback function(err, accepted)
+  done // on done callback function(err, result)
+) => {
+  const len = items.length;
+  let i = 0;
+  done = common.cb(done);
+
+  function next() {
+    if (i === len) return done(null, false);
+    fn(items[i], (err, accepted) => {
+      if (err) return done(err);
+      if (accepted) return done(null, true);
+      i++;
+      next();
+    });
+  }
+
+  if (len > 0) next();
+  else done(null, false);
+};
+
+module.exports = {
+  map,
+  filter,
+  reduce,
+  each,
+  series,
+  find,
+  every,
+  some
 };
