+++ test/fp.of.js
@@ -5,7 +5,7 @@ const metasync = require('..');
 
 tap.test('of test', (test) => {
   const args = [1, 2, 3, 4, 5];
-  metasync.monad.of(...args)((err, ...argsCb) => {
+  metasync.of(...args)((err, ...argsCb) => {
     test.error(err);
     test.strictSame(args, argsCb);
     test.end();
